requirements.txt
anyio==3.7.0
APScheduler==3.10.1
certifi==2023.5.7
charset-normalizer==3.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
joblib==1.2.0
navec==0.10.0
numpy==1.24.3
pandas==2.0.2
psycopg2==2.9.6
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
